Collapse approval projects into one
and update to .NET 4.6.1 and xUnit 2.2, for consistency with other
projects
