+++ FakeItEasy.sln
@@ -123,8 +123,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeItEasy.Tests.Approval",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeItEasy", "src\FakeItEasy\FakeItEasy.csproj", "{74165807-676F-4464-B607-8770C13E8991}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeItEasy.Tests.Approval.netstd", "tests\FakeItEasy.Tests.Approval.netstd\FakeItEasy.Tests.Approval.netstd.csproj", "{BC38E29E-2915-4E3E-BC12-69226747E269}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeItEasy.Analyzer.VisualBasic", "src\FakeItEasy.Analyzer.VisualBasic\FakeItEasy.Analyzer.VisualBasic.csproj", "{7BAEB7DC-635D-425B-94AB-826AB65FBBD7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FakeItEasy.Analyzer.VisualBasic.Tests", "tests\FakeItEasy.Analyzer.VisualBasic.Tests\FakeItEasy.Analyzer.VisualBasic.Tests.csproj", "{C37E7B74-DCA9-47FF-BAA1-EBC2ECB2C72F}"
@@ -191,10 +189,6 @@ Global
 		{74165807-676F-4464-B607-8770C13E8991}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{74165807-676F-4464-B607-8770C13E8991}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{74165807-676F-4464-B607-8770C13E8991}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BC38E29E-2915-4E3E-BC12-69226747E269}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BC38E29E-2915-4E3E-BC12-69226747E269}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BC38E29E-2915-4E3E-BC12-69226747E269}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BC38E29E-2915-4E3E-BC12-69226747E269}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7BAEB7DC-635D-425B-94AB-826AB65FBBD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7BAEB7DC-635D-425B-94AB-826AB65FBBD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7BAEB7DC-635D-425B-94AB-826AB65FBBD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -224,7 +218,6 @@ Global
 		{9D4A41F1-D444-4244-AAAC-F610D9977B2F} = {AD790E7D-D664-4B5F-8BE8-01507A2A9F3B}
 		{42A92357-2A87-4B87-A591-CA78B6CE4461} = {9D4A41F1-D444-4244-AAAC-F610D9977B2F}
 		{74165807-676F-4464-B607-8770C13E8991} = {59466C4D-C407-4DD9-B77B-D99162B9FC7C}
-		{BC38E29E-2915-4E3E-BC12-69226747E269} = {9D4A41F1-D444-4244-AAAC-F610D9977B2F}
 		{7BAEB7DC-635D-425B-94AB-826AB65FBBD7} = {59466C4D-C407-4DD9-B77B-D99162B9FC7C}
 		{C37E7B74-DCA9-47FF-BAA1-EBC2ECB2C72F} = {DC5EAE30-96E3-4D38-B3C5-33E587C9FB76}
 		{C4766C64-39E1-4023-A173-615607C40AD5} = {59466C4D-C407-4DD9-B77B-D99162B9FC7C}
