--- tests/FakeItEasy.Tests.Approval.netstd/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ApprovalTests" version="3.0.11" targetFramework="net45" />
-  <package id="ApprovalUtilities" version="3.0.11" targetFramework="net45" />
-  <package id="PublicApiGenerator" version="4.0.1" targetFramework="net45" />
-  <package id="StyleCop.Analyzers" version="1.0.0" targetFramework="net45" developmentDependency="true" />
-  <package id="xunit.abstractions" version="2.0.0" targetFramework="net45" />
-  <package id="xunit.extensibility.core" version="2.1.0" targetFramework="net45" />
-  <package id="xunit.extensibility.execution" version="2.1.0" targetFramework="net45" />
-  <package id="xunit.runner.visualstudio" version="2.1.0" targetFramework="net45" />
-</packages>
