+++ tests/FakeItEasy.Tests.Approval/ApiApproval.ApproveApiNetStd.approved.txt
